==> notebooks/3_atlas_extension/01_final_data_cleaning_and_prep_for_scArches.ipynb <==
## Preparation of HLCA extension datasets for mapping to the HLCA core:

We'll use the data files that have been prepared in the notebooks from the HLA_extension_data_preprocessing folder as a basis for the HLCA extension. Here we perform final checks, as well as further cleaning and harmonization of metadata.

### TO DO: ADAPT NOTEBOOK SO THAT IT INCLUDES THE HLCA CORE (FOR BOTTOM PART)<br>

### Import modules, set paths:

import scanpy as sc
import os
import pandas as pd
import sys

sys.path.append("../../scripts/")
import reference_based_harmonizing
import preprocessing

for nice formatting of code (not necessary):

%load_ext lab_black

set paths:

dir_data = "../../data/HLCA_extended/extension_datasets/ready/subsetted/"
path_output_atlas_extension = (
    "../../data/HLCA_extended/HLCA_extended/extension_data_pooled_raw_2000genes.h5ad"
)
path_cell_label_harmonization = "../../supporting_files/metadata_harmonization/HLCA_cell_type_reference_mapping_20220926.csv"

get data file names to import:

files_to_import = sorted(os.listdir(dir_data))

### Read in all files and check if integers:

read in prepared anndata objects:

verbose = True
query_adatas = dict()
for dataset in files_to_import:
    dataset_name = dataset.split("_sub.h5ad")[0].split(".h5ad")[0]
    query_adatas[dataset_name] = sc.read(f"../query_datasets/ready/subsetted/{dataset}")
    if dataset_name == "lukassen":
        if verbose:
            print(
                "rounding counts for lukassen dataset, the counts had slight rounding errors leading to non-integer values"
            )
            print("pre:")
            print(query_adatas[dataset_name].X[:10, :10].toarray())
        query_adatas[dataset_name].X = round(query_adatas[dataset_name].X)
        if verbose:
            print("post:")
            print(query_adatas[dataset_name].X[:10, :10].toarray())
    if verbose and dataset_name != "lukassen":
        print(dataset_name, query_adatas[dataset_name].shape)
        print(query_adatas[dataset_name].X[:10, :10].toarray())

### Add, filter and clean metadata:

rename and filter adata.obs variables:

obs_to_rename = {
    "cond": "condition",
    "anatomical_region_fine": "anatomical_region_detailed",
    "orig_celltype_ann": "original_celltype_ann",
    "bmi": "BMI",
    "subject_id": "subject_ID",
}

obs_to_keep = [
    "sample",
    "condition",
    "study",
    "dataset",
    "original_celltype_ann",
    "sample",
    "age",
    "sex",
    "disease",
    "subject_ID",
    "sample_type",
    "ethnicity",
    "BMI",
    "smoking_status",
    "anatomical_region_coarse",
    "anatomical_region_detailed",
    "anatomical_region_level_1",
    "genome",
    "single_cell_platform",
]

verbose = False
for ds, ad in query_adatas.items():
    if verbose:
        print("dataset:", ds)
    obs_renamer = {obscol: obscol for obscol in ad.obs.columns}
    for old_name, new_name in obs_to_rename.items():
        obs_renamer[old_name] = new_name
    ad.obs.rename(columns=obs_renamer, inplace=True)
    # remove all columns that are not in our to-keep list:
    col_to_remove = [col for col in ad.obs.columns if col not in obs_to_keep]
    if verbose:
        print("columns to remove:", col_to_remove)
    ad.obs.drop(columns=col_to_remove, inplace=True)
    # remove all var columns:
    ad.var.drop(columns=ad.var.columns, inplace=True)
    # store:
    query_adatas[ds] = ad

add sample type info, single cell protocol info, and single-cell vs single nucleus info. We will later add the information to datasets that are mixed in these respects.

sample_type_info = {
    "adams": "MIXED",  # normal: donor lung, diseased: lung explant,
    "barbry": "MIXED_DONE",  # already annotated in object
    "bharat": "autopsy",
    "carraro_cff": "lung_explant",  # explant
    "carraro_csmc": "lung_explant",  # explant
    "carraro_ucla": "lung_explant",  # explant
    "delorey_cryo": "autopsy_cryopreserved",
    "delorey_fresh": "autopsy",
    "delorey_nuclei": "autopsy",
    "duong": "lung_explant",  # explant
    "grant_cryo": "balf_cryopreserved",
    "grant_fresh": "balf",
    "guo": "lung_explant",  # explant
    "haberman": "lung_explant",  # explant
    "lambrechts": "surgical_resection",
    "laughney": "surgical_resection",
    "liao": "balf",
    "lukassen": "surgical_resection",
    "mayr": "MIXED",  # healthy: surgical resection, ILD: lung explant
    "meyer": "donor_lung",
    "mould": "balf",
    "ordovasmontanes": "scraping",
    "reyfman_disease": "MIXED",  # explanted lung and cryobiopsy
    "schiller_discovair": "surgical_resection",
    "szabo": "donor_lung",
    "tsukui": "donor_lung",  # rejected donors
    "valenzi": "lung_explant",  # explant
    "wang_sub_batch1": "donor_lung_cryopreserved",
    "wang_sub_batch2": "donor_lung_cryopreserved",
    "wang_sub_batch3": "donor_lung_cryopreserved",
    "wang_sub_batch4": "donor_lung_cryopreserved",
    "wouters": "balf",
}

cell_vs_nucleus = {
    "adams": "cells",
    "barbry": "cells",
    "bharat": "cells",
    "carraro_cff": "cells",
    "carraro_csmc": "cells",
    "carraro_ucla": "cells",
    "delorey_cryo": "cells",
    "delorey_fresh": "cells",
    "delorey_nuclei": "nuclei",
    "duong": "nuclei",
    "grant_cryo": "cells",
    "grant_fresh": "cells",
    "guo": "cells",
    "haberman": "cells",
    "lambrechts": "cells",
    "laughney": "cells",
    "liao": "cells",
    "lukassen": "nuclei",
    "mayr": "cells",
    "meyer": "cells",
    "mould": "cells",
    "ordovasmontanes": "cells",
    "reyfman_disease": "cells",
    "schiller_discovair": "cells",
    "szabo": "cells",
    "tsukui": "cells",
    "valenzi": "cells",
    "wang_sub_batch1": "nuclei",
    "wang_sub_batch2": "nuclei",
    "wang_sub_batch3": "nuclei",
    "wang_sub_batch4": "nuclei",
    "wouters": "cells",
    "Meyer_2021":"cells",
    "MeyerNikolic_unpubl":"cells",
    "Tata_unpubl":"cells",
    "Schultze_unpubl":"cells"
    
}

This information is further refined at the bottom of this notebooks.

protocols = {
    "adams": "10X v2",
    "barbry": "10X v3 3'",
    "bharat": "10X v3",
    "carraro_cff": "10X",
    "carraro_csmc": "10X",
    "carraro_ucla": "Drop-Seq",
    "delorey_cryo": "10X v1.1 5'",
    "delorey_fresh": "10X v1.1 5'",
    "delorey_nuclei": "10X v3 3'",
    "duong": "10X v3",
    "grant_cryo": "10X v1 5'",
    "grant_fresh": "10X v1 5'",
    "guo": "10X v2",
    "haberman": "10X",
    "lambrechts": "10X",  # mixed
    "laughney": "10X v2",
    "liao": "10X",  # mixed v2 and v3
    "lukassen": "10X v2",
    "mayr": "Drop-Seq",
    "meyer": "MIXED_DONE",
    "mould": "10X v2",
    "ordovasmontanes": "Seq-Well",
    "reyfman_disease": "10X v2 3'",
    "schiller_discovair": "10X v3 3'",
    "szabo": "10X 3'",
    "tsukui": "10X v2 3'",
    "valenzi": "10X v2",
    "wang_sub_batch1": "10X v3",
    "wang_sub_batch2": "10X v3",
    "wang_sub_batch3": "10X v3",
    "wang_sub_batch4": "10X v3",
    "wouters": "10X 5'",
}

for ds_name, ds in query_adatas.items():
    sample_type = sample_type_info[ds_name]
    cells_or_nuclei = cell_vs_nucleus[ds_name]
    protocol = protocols[ds_name]
    if not sample_type.startswith("MIXED"):
        ds.obs["sample_type"] = sample_type
    else:
        print("Add sample type info later for", ds_name)
    if not cells_or_nuclei == "unknown":
        ds.obs["cells_or_nuclei"] = cells_or_nuclei
    else:
        print("cells or nuclei unknown for", ds_name)
        ds.obs["cells_or_nuclei"] = "unknown"
    if protocol == "unknown":
        print("single cell protocol unknown for", ds_name)
        ds.obs["single_cell_platform"] = "unknown"
    elif protocol.startswith("MIXED"):
        print("mixed single cell for", ds_name)
    else:
        ds.obs["single_cell_platform"] = protocol

    query_adatas[ds_name] = ds

Add dataset specific info/corrections:

Cancer type for lambrechts data:

lambrechts_subject_to_condition = {
    "lambrechts_1": "Squamous Cell Carcinoma",
    "lambrechts_2": "Squamous Cell Carcinoma",
    "lambrechts_3": "Adenocarcinoma",
    "lambrechts_4": "Adenocarcinoma",
    "lambrechts_5": "Large Cell Carcinoma",
    "lambrechts_6": "Adenocarcinoma",
    "lambrechts_7": "Squamous Cell Carcinoma",
    "lambrechts_8": "Pleiomorphic Carcinoma",
}

verbose = False
for ds_name, ds in query_adatas.items():
    if verbose:
        print(ds_name)
    # add sample type info to studies with mixed sample types:
    if ds_name == "adams":
        adams_cond_to_st = {
            "Control": "donor_lung",
            "IPF": "lung_explant",
            "COPD": "lung_explant",
        }
        ds.obs["sample_type"] = ds.obs.condition.map(adams_cond_to_st)
    if ds_name == "mayr":
        mayr_cond_to_st = {
            "control donor": "surgical_resection",
            "endstage lung fibrosis": "lung_explant",
        }
        ds.obs["sample_type"] = ds.obs.condition.map(mayr_cond_to_st)
    if ds_name == "reyfman_disease":
        reyfman_cond_to_st = {
            cond: "lung_explant" for cond in ds.obs.condition.unique()
        }
        reyfman_cond_to_st["Cryobiopsy"] = "biopsy_cryopreserved"
        ds.obs["sample_type"] = ds.obs.condition.map(reyfman_cond_to_st)
    # remove duplicated cells in mould/janssen dataset (preprocessing mistake)
    if ds_name == "mould":
        ds = ds[~pd.isnull(ds.obs.subject_ID), :].copy()
    # add dataset and study info for duong
    if ds_name == "duong":
        ds.obs["dataset"] = "Duong_lungMAP_unpubl"  # TO DO WAIT FOR MALTE'S RESPONSE
        ds.obs["study"] = "Duong_lungMAP_unpubl"
    if ds_name == "meyer":
        ds.obs["condition"] = "Healthy"  # healthy lung from donors
    # remove duplicated condition column for tsukui data
    if ds_name == "tsukui":
        ds.obs = ds.obs.loc[:, ~ds.obs.columns.duplicated()].copy()
    # correct condition for guo (now only set to "disease"):
    if ds_name == "guo":
        ds.obs["condition"] = "Lymphangioleiomyomatosis"
    # add condition healthy to wang datasets:
    if ds_name.startswith("wang"):
        ds.obs["condition"] = "Healthy"
    # add cancer type info for lambrechts data:
    if ds_name == "lambrechts":
        ds.obs["condition"] = ds.obs.subject_ID.map(lambrechts_subject_to_condition)
    # make sample names unique (not just numbers) for the following datasets:
    if ds_name in [
        "barbry",
        "duong",
        "grant_cryo",
        "grant_fresh",
        "laughney",
        "reyfman",
    ]:
        study = ds.obs.study.unique()[0]
        # check if not already corrected:
        if not str(ds.obs["sample"][0]).startswith(study):
            ds.obs["sample"] = [
                f"{study}_sample_{s_number}" for s_number in ds.obs["sample"]
            ]
    # make subject names unique (not just numbers) for the following datasets:
    if ds_name in ["ordovasmontanes", "grant_cryo", "grant_fresh"]:
        study = ds.obs.study.unique()[0]
        if not str(ds.obs["subject_ID"][0]).startswith(study):
            ds.obs["subject_ID"] = [
                f"{study}_subject_{s_number}" for s_number in ds.obs["subject_ID"]
            ]
    # remove spaces from subject names:
    ds.obs["subject_ID"] = [sid.replace(" ", "_") for sid in ds.obs["subject_ID"]]

    # STORE RESULT:
    query_adatas[ds_name] = ds

### Merge all datasets into one object:

merge datasets into one object:

adata = sc.AnnData.concatenate(
    *query_adatas.values(),
    join="outer",
    batch_key=None,
    batch_categories=list(query_adatas.keys()),
    index_unique="_"
)
print(adata.shape)

adata

### Further metadata cleaning:

renaming of conditions, studies, datasets, sex, ethnicity:

condition_renaming = {
    "CO": "Control", # should I change this to healthy? 
    "healthy": "Healthy",
    "TUMOR": "Lung adenocarcinoma",
    "NOR": "Tumor adjacent normal",
    "control": "Control",
    "moderate COVID-19": "COVID-19 moderate",
    "severe COVID-19": "COVID-19 severe",
    "smoking": "Healthy",
    "EAA": "HP",
    "control donor": "Control",
    "endstage lung fibrosis": "Endstage lung fibrosis",
    "chronic rhinitis": "Chronic rhinitis",
    "respiratory system disease": "Respiratory system disease",
    "SSc-ILD": "Systematic sclerosis-associated ILD",
    "SSc": "Systematic sclerosis-associated ILD",
    "Myositis-ILD": "Myositis-associated ILD",
    "Cryobiopsy": "IPF",  # this is a single sample from Reyfman et al.
    "stimulated": "anti-CD3 anti-CD28 antibody stimulated",
    "SCD": "Scleroderma",
    "NML": "Healthy",
    "non-COVID pneumonia": "Pneumonia non-COVID",
}  # IPF, Control, COPD, CF, ILD, Sarcoidosis, NSIP, cHP

study_renamer = {
    "BarbyUnpublished": "Barbry_unpubl",
    "Kaminski2020": "Kaminski_2020",
    "Budinger2021": "Budinger_2020",
    "Gomperts2021": "Gomperts_2021",
    "regev2021": "Regev_2021",
    "Wunderink2021": "Wunderink_2021",
    "Guo2020": "Xu_2020",
    "Banovich2020": "Banovich_Kropski_2020",
    "PeerMassague2020": "Peer_Massague_2020",
    "Zhang2021": "Zhang_2021",
    "Eils2020": "Eils_2020",
    "Schiller2020": "Schiller_2020",
    "Janssen2020": "Janssen_2020",
    "Shalek2018": "Shalek_2018",
    "MisharinBudinger2019": "Misharin_Budinger_2018",
    "Schiller2021": "Schiller_2021",
    "Sims2019": "Sims_2019",
    "Sheppard2020": "Sheppard_2020",
    "Lafyatis2019": "Lafyatis_2019",
    "Sun2020": "Sun_2020",
    "Lambrechts2021": "Lambrechts_2021",
}

dataset_renamer = {
    "BarbyUnpublished": "Barbry_unpubl",
    "Kaminski2020": "Kaminski_2020",
    "Budinger2021": "Budinger_2020",
    "Gomperts2021_CFF": "Gomperts_2021_CFF",
    "Gomperts2021_CSMC": "Gomperts_2021_CSMC",
    "Wunderink2021_cryo": "Wunderink_2021_cryo",
    "Wunderink2021_fresh": "Wunderink_2021_fresh",
    "Guo2020_LAM1/3": "Xu_2020_LAM1_3",
    "Banovich2020": "Banovich_Kropski_2020",
    "PeerMassague2020": "Peer_Massague_2020",
    "Zhang2021": "Zhang_2021",
    "Eils2020": "Eils_2020",
    "Schiller2020": "Schiller_2020",
    "Janssen2020": "Janssen_2020",
    "Shalek2018": "Shalek_2018",
    "MisharinBudinger2019_disease": "Misharin_Budinger_2018",
    "Schiller2021": "Schiller_2021",
    "Sims2019": "Sims_2019",
    "Sheppard2020": "Sheppard_2020",
    "Lafyatis2019": "Lafyatis_2019",
    "Sun2020_batch1": "Sun_2020_batch1",
    "Sun2020_batch2": "Sun_2020_batch2",
    "Sun2020_batch3": "Sun_2020_batch3",
    "Sun2020_batch4": "Sun_2020_batch4",
    "Lambrechts2021": "Lambrechts_2021",
    "KULeuven_Thienpont_2018Lambrechts_v1" = "Thienpont_2018_10Xv1",
    "KULeuven_Thienpont_2018Lambrechts_v2" = "Thienpont_2018_10Xv2"
    
}

sex_renamer = {
    'F': 'female',
    'M': 'male',
    'male': 'male',
    'Female': 'female',
    'Male': 'male',
    'female': 'female'
}

ethn_renamer = {
    "Latina":"latino",
    "None":np.nan,
    "nan":np.nan
}

age_renamer = {"37yo": "37", "nan": np.nan}

Now include all values that stay the same during the remapping:

condition_renaming_full = {cond: cond for cond in adata.obs.condition.unique()}
for cond_old, cond_new in condition_renaming.items():
    condition_renaming_full[cond_old] = cond_new
study_renaming_full = {st: st for st in adata.obs.study.unique()}
for st_old, st_new in study_renamer.items():
    study_renaming_full[st_old] = st_new
dataset_renaming_full = {ds: ds for ds in adata.obs.dataset.unique()}
for ds_old, ds_new in dataset_renamer.items():
    dataset_renaming_full[ds_old] = ds_new
sex_renaming_full = {sex:sex for sex in adata.obs.sex.unique()}
for sex_old, sex_new in sex_renamer.items():
    sex_renaming_full[sex_old] = sex_new
ethn_renaming_full = {ethn:ethn for ethn in adata.obs.ethnicity.unique()}
for ethn_old, ethn_new in ethn_renamer.items():
    ethn_renaming_full[ethn_old] = ethn_new
age_renaming_full = {age:age for age in adata.obs.age.unique()}
for age_old, age_new in age_renamer.items():
    age_renaming_full[age_old] = age_new

now map annotations:

adata.obs["condition"] = adata.obs.condition.map(condition_renaming_full)
adata.obs["study"] = adata.obs.study.map(study_renaming_full)
adata.obs["dataset"] = adata.obs.dataset.map(dataset_renaming_full)
adata.obs["sex"] = adata.obs.sex.map(sex_renaming_full)
adata.obs["ethnicity"] = adata.obs.ethnicity.map(ethn_renaming_full)
adata.obs["age"] = adata.obs.age.map(age_renaming_full)


convert age to float:

adata.obs.age = [np.float(age) for age in adata.obs.age]

### Sanity checks:

check if samples are unique to one subject, dataset and study:

data_per_sample = adata.obs.groupby("sample").agg(
    {"dataset": "nunique", "subject_ID": "nunique", "study": "nunique"}
)

this should all be 0:

(data_per_sample > 1).sum(axis=0)

same for subjects:

data_per_subject = adata.obs.groupby("subject_ID").agg(
    {"dataset": "nunique", "study": "nunique"}
)

3 subjects are present in two regev datasets from the same study, so this should be 3 for dataset and 0 for study

(data_per_subject > 1).sum(axis=0)

check if all obs are fully annotated:

for col in adata.obs.columns:
    print(col)
    print("Fully annotated?", ~pd.isnull(adata.obs[col]).any())
    print("categories:", adata.obs[col].unique())
    print("\n\n")

## Add harmonized annotations where possible:

adata = sc.read_h5ad(path_output_atlas_extension)

path_output_atlas_extension

adata

List the datasets  for which we have mapped original annotations to our hierarchical cell type reference; for these we will do the actual harmonizing mapping:

datasets_harmonized = [
    "Banovich_Kropski_2020",
    "Banovich_Kropski_2020_disease",  # temporarily name "Banovich_Kropski_2020" this way: it was annotated independently with different terminology
    "Duong_lungMAP_unpubl",
    "Eils_2020",
    "Kaminski_2020",
    "Meyer_2021",
    "Sheppard_2020",
    "Sun_2020",
    "Thienpont_2018",
]

Subset the data to the datasets listed above:

adata_to_harmonize = adata[adata.obs.study.isin(datasets_harmonized), :].copy()

temporarily name "Banovich_Kropski_2020" "Banovich_Kropski_2020_disease": it was annotated independently with different terminology, and the mapping is listed under that study name.

adata_to_harmonize.obs.study = adata_to_harmonize.obs.study.tolist()

adata_to_harmonize.obs.loc[
    adata.obs.study == "Banovich_Kropski_2020", "study"
] = "Banovich_Kropski_2020_disease"

Now perform the mapping.

First, import the .csv that contains the translations:

harmonizing_df = reference_based_harmonizing.load_harmonizing_table(
    path_cell_label_harmonization
)

Create a dataframe that contains each cell type name from the consensus ontology as indices, with their matching annotations at the other levels. This will simplify mapping:

consensus_df = reference_based_harmonizing.create_consensus_table(harmonizing_df)

create a dataframe that for each original celltype annotation (from all datasets pooled) provides the translation to the consensus reference at all levels:

celltype_translation_df = (
    reference_based_harmonizing.create_orig_ann_to_consensus_translation_df(
        adata_to_harmonize, consensus_df, harmonizing_df, verbose=False
    )
)

now translate the original annotations to the consensus in your AnnData:

adata_to_harmonize = reference_based_harmonizing.consensus_annotate_anndata(
    adata_to_harmonize, celltype_translation_df, verbose=True
)

Check if anything was mapped to the category "unicorns and artifacts" (cells annotated as low quality, doublets etc.). If this is not zero, remove the cells!

print(
    "Number of unicorns and artifact cells to remove:",
    sum(adata_to_harmonize.obs.original_ann_level_1 == "Unicorns and artifacts"),
)

add "clean" annotations (with "None" for cells that do not have annotation at that level):

adata_to_harmonize = reference_based_harmonizing.add_clean_annotation(
    adata_to_harmonize, input_ann_type="original"
)

Now copy results to full adata object:

for col in adata_to_harmonize.obs.columns:
    if col.startswith("original_ann_level"):
        adata.obs[col] = list(adata.obs[col])  # so that we can add new categories
        adata.obs.loc[adata_to_harmonize.obs.index, col] = adata_to_harmonize.obs[col]

## Cleaning, round 2:

Note that this includes the HLCA core data, might have to update the notebook, as above is only extension?

### add study or dataset specific information that was missing:

study_scp_mapper = {
    "Budinger_2020": "10X 3' v3",
    "Duong_lungMAP_unpubl": "10X 3' v3",
    "Eils_2020": "10X 3' v2",
    "Janssen_2020": "10X 3' v2",
    "Kaminski_2020": "10X 3' v2",
    "Lafyatis_2019": "10X 3' v2",
    "Lambrechts_2021":"10X 5' v1", 
    "Peer_Massague_2020": "10X 3' v2",
    "Sims_2019": "10X 3' v2",  
    "Sun_2020": "10X 3' v3",
    "Thienpont_2018": "10X 3'",  # both v1 and v2
    "Xu_2020": "10X 3' v2",
    "Zhang_2021": "10X 3'",  # both v2 and v3
}

study_cvn_mapper = {
    "MeyerNikolic_unpubl": "cells",
    "Meyer_2021": "cells",
    "Tata_unpubl": "cells",
    "Schultze_unpubl":"cells"
}

study_cond_mapper = {
    "Barbry_Leroy_2020": "Healthy",
    "Eils_2020": "Healthy (tumor adjacent)",
    "Krasnow_2020": "Healthy (tumor adjacent)",
    "Jain_Misharin_2021": "Healthy",
    "Lafyatis_Rojas_2019": "Healthy",
    "Misharin_2021": "Healthy",
    "Schiller_2021": "Healthy (tumor adjacent)",
}

study_samt_mapper = {"Duong_lungMAP_unpubl": "donor_lung", "Tata_unpubl": "autopsy"}

ds_scp_mapper = {"Gomperts_2021_CFF": "10X 3' v2", "Gomperts_2021_CSMC": "10X 3' v2"}

adata = preprocessing.update_adata(
    adata, "study", "single_cell_platform", study_scp_mapper
)

adata = preprocessing.update_adata(
    adata, "study", "cells_or_nuclei", study_cvn_mapper
)

adata = preprocessing.update_adata(
    adata, "study", "condition", study_cond_mapper
)

adata = preprocessing.update_adata(
    adata, "study", "sample_type", study_samt_mapper
)

adata = preprocessing.update_adata(
    adata, "dataset", "single_cell_platform", ds_scp_mapper
)

additional manual changes:
- Gomperts_2021 controls are (rejected) donor lungs
- Misharin_Budinger_2918 condition "nan" is healthy
- Schiller_2020 control is tumor adjacent healthy
- Shalek 2018 - Set healthy controls (donors 358, 360, 362) to healthy instead of chronic rhinitis. Set respiratory system disease to Chronic rhinitis. (Based on communication with Amanda Oliver who looked into this)
- Zhang 2021 control is healthy
- Schultze_unpubl is tumor adjacent normal, surgical resections
- Seibold_2020 is healthy
- Teichman_meyer and Meyer are healthy 
- Banovich Kropski 10X is 10X 5' (not sure about version) and condition nans are healthy
- Kaminski_2020 Control is healthy

# Gomperts
adata.obs.loc[
    [
        st == "Gomperts_2021" and cond == "Control"
        for st, cond in zip(adata.obs.study, adata.obs.condition)
    ],
    "condition",
] = "Healthy"
adata.obs.loc[
    [
        st == "Gomperts_2021" and cond == "Control"
        for st, cond in zip(adata.obs.study, adata.obs.condition)
    ],
    "sample_type",
] = "donor_lung"
# Misharin_Budinger
adata.obs.loc[
    [
        st == "Misharin_Budinger_2018" and cond == "nan"
        for st, cond in zip(adata.obs.study, adata.obs.condition)
    ],
    "condition",
] = "Healthy"
# Schiller
adata.obs.loc[
    [
        st == "Schiller_2020" and cond == "Control"
        for st, cond in zip(adata.obs.study, adata.obs.condition)
    ],
    "condition",
] = "Healthy (tumor adjacent)"
# Shalek
shalek_healthy_subjects = [
    "Shalek2018_subject_358",
    "Shalek2018_subject_360",
    "Shalek2018_subject_362",
]
adata.obs.loc[
    [
        st == "Shalek_2018" and cond == "Respiratory system disease"
        for st, cond in zip(adata.obs.study, adata.obs.condition)
    ],
    "condition",
] = "Chronic rhinitis"
adata.obs.loc[
    [
        st == "Shalek_2018" and subj in shalek_healthy_subjects
        for st, subj in zip(adata.obs.study, adata.obs.subject_ID)
    ],
    "condition",
] = "Healthy"
# Zhang
adata.obs.loc[
    [
        st == "Zhang_2021" and cond == "Control"
        for st, cond in zip(adata.obs.study, adata.obs.condition)
    ],
    "condition",
] = "Healthy"
# Schultze
adata.obs.loc[
    [st == "Schultze_unpubl" for st in adata.obs.study], "condition"
] = "Healthy (tumor adjacent)"
adata.obs.loc[
    [st == "Schultze_unpubl" for st in adata.obs.study], "sample_type"
] = "surgical_resection"
# Seibold
adata.obs.loc[
    [st == "Seibold_2020" for st in adata.obs.study], "condition"
] = "Healthy"
# Teichmann_meyer
adata.obs.loc[
    [st == "Teichmann_Meyer_2019" for st in adata.obs.study], "condition"
] = "Healthy"
# Meyer
adata.obs.loc[
    [st == "Meyer_2019" for st in adata.obs.study], "condition"
] = "Healthy"
# Banovich_Kropski
adata.obs.loc[
    [
        st == "Banovich_Kropski_2020" and scp == "10X"
        for st, scp in zip(adata.obs.study, adata.obs.single_cell_platform)
    ],
    "single_cell_platform",
] = "10X 5'"
adata.obs.condition = adata.obs.condition.tolist()
adata.obs.loc[
    [
        st == "Banovich_Kropski_2020" and cond == "nan"
        for st, cond in zip(adata.obs.study, adata.obs.condition)
    ],
    "condition",
] = "Healthy"
adata.obs.condition = pd.Categorical(adata.obs.condition)
# Kaminski
adata.obs.loc[
    [
        st == "Kaminski_2020" and cond == "Control"
        for st, cond in zip(adata.obs.study, adata.obs.condition)
    ],
    "condition",
] = "Healthy"

Harmonize terminology:

scp_cleaner = {
    "10X v1 5'": "10X 5' v1",
    "10X v1.1 5'": "10X 5' v1",
    "10X v2 3'": "10X 3' v2",
    "10X v3 3'": "10X 3' v3",
    "10x_3'_v1": "10X 3' v1",
    "10x_3'_v2": "10X 3' v2",
    "10x_3'_v3": "10X 3' v3",
    "10x_5'_v1": "10X 5' v1",
    "10x_5'_v1.1": "10X 5' v1",
    "10x_5'_v2": "10X 5' v2",
    "Seq-well": "Seq-Well",
}

adata = preprocessing.update_adata(
    adata, "single_cell_platform", "single_cell_platform", scp_cleaner
)

sorted(adata.obs.single_cell_platform.unique())

condition_cleaner = {
    "Adenocarcinoma": "Lung adenocarcinoma",
    "healthy": "Healthy",
    "Tumor adjacent normal": "Healthy (tumor adjacent)",
    "Systematic sclerosis-associated ILD": "Systemic sclerosis-associated ILD",
    "Scleroderma": "Systemic sclerosis-associated ILD",
    "Endstage lung fibrosis": "End-stage lung fibrosis, unknown etiology",
    "cHP": "HP",
}

adata = preprocessing.update_adata(adata, "condition", "condition", condition_cleaner)

sorted(adata.obs.condition.unique())

update study names:

study_cleaner = {"MeyerNikolic_unpubl": "Meyer_Nikolic_2022"}

adata = preprocessing.update_adata(adata, "study", "study", study_cleaner)

Check results:

study_info_new = adata.obs.groupby("study").agg(
    {
        "dataset": "nunique",
        "study": "count",
        "subject_ID": "nunique",
        "sample": "nunique",
        "sample_type": "nunique",
        "subject_type": "nunique",
        "single_cell_platform": "nunique",
        "condition": "nunique",
        "cells_or_nuclei": "nunique",
    }
)
for cov in ["sample_type", "subject_type", "single_cell_platform", "condition"]:
    study_info_new[f"{cov}_set"] = [
        adata.obs.loc[adata.obs.study == study, cov].unique().tolist()
        for study in study_info_new.index
    ]

study_info_new

## Store result:

adata

path_output_atlas_extension

adata.write(path_output_atlas_extension)

copy updated annotations to the embedding h5ad:

adata_ext = sc.read_h5ad(
    "../../data/HLCA_extended/HLCA_extended/HLCA_extended_scarches_emb.h5ad"
)

adata_ext

for col in adata.obs.columns:
    if col.startswith("original_ann"):
        adata_ext.obs[col] = adata_ext.obs[col].tolist()
        adata_ext.obs.loc[adata.obs.index, col] = adata.obs[col]

adata_ext

adata_ext.write(
    "../../data/HLCA_extended/HLCA_extended/HLCA_extended_scarches_emb.h5ad"
)
