==> notebooks/3_atlas_extension/12_full-matrix_features_digestion.ipynb <==
## Gene names processing

In this notebook we are digesting idtrack's output. We take the dictionaries and create a table of genes-ensembl pairs to determine what the final gene name will be. In the process we also find repeated gene names that will be merged, per dataset.

__author__ = "Ciro Ramírez-Suástegui"

__copyright__ = "Copyright 2022-11-09, Helmholtz Zentrum Muenchen"

__license__ = "GPL"

__version__ = "0.0.9"

__email__ = "ciro.suastegui@helmholtz-muenchen.de, ksuasteguic@gmail.com"

__status__ = "Prototype"

#### Structure

* [Global variables and paths](#bullet1)
* [Loading data](#bullet2)
* [Pre-processing](#bullet3)
* [Main](#bullet4)
* [Conclusions](#bullet5)
* [Save](#bullet6)

### Environment setup

%load_ext autoreload
%autoreload 2
import importlib

spam_spec = importlib.util.find_spec("lab_black")
if spam_spec is not None:
    %load_ext lab_black

# basic modules
import warnings, os, re
import time, sys

warnings.simplefilter(action="ignore", category=FutureWarning)
warnings.simplefilter(action="ignore", category=UserWarning)

# in-house/developing modules
# tools modules
import scanpy as sc
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import itertools
from collections import Counter
from tqdm import tqdm
import pickle

sc.logging.print_versions()

print("Environment:", re.sub(".os.py", "", os.__file__))
parentpath = lambda _path, n: os.sep.join(_path.split(os.sep)[:-n])
print("Working at:", re.sub(parentpath(os.getcwd(), 2), "", os.getcwd()))

def make_symbol(x):
    return re.sub(
        "^ENSG[0-9]{10,}\.[0-9]{1,}_|_ENSG[0-9]{10,}",  # "ENSG.*_|_ENSG[0-9]{10,}",
        "",
        re.sub(  # some artifacts from idtrack returns the mapping structure
            ".*\n",
            "",
            re.sub(  # which contains several names
                ".* {1,}",
                "",
                re.sub(
                    "\nName.*",
                    "",
                    x,
                ),
            ),
        ),
    )

### Global variables and paths <a class="anchor" id="bullet1"></a>

feats_dir = "../../data/HLCA_extended/extension_datasets/features/"
feats_inp = (
    "../../data/HLCA_extended/extension_datasets/features/all_update-with-core.pk"
)
adata_split = {
    "HLCA": {
        "study": [
            "Nawijn_2021_HLCA",
            "Barbry_Leroy_2020_HLCA",
            "Meyer_2019_HLCA",
            "Banovich_Kropski_2020_HLCA",
            "Seibold_2020_HLCA",
            "Jain_Misharin_2021_HLCA",
            "Teichmann_Meyer_2019_HLCA",
            "Misharin_2021_HLCA",
            "Lafyatis_Rojas_2019_HLCA",
            "Krasnow_2020_HLCA",
            "Misharin_Budinger_2018_HLCA",
        ]
    }
}

### Loading data <a class="anchor" id="bullet2"></a>

results_core = pd.read_pickle(
    "/lustre/groups/ml01/workspace/kemal.inecik/idtrack_temp/results_for_hlca_core_only_HGNC Symbol_20221110-094335.pk"
)
results_idtrack = pd.read_pickle(
    "/lustre/groups/ml01/workspace/kemal.inecik/idtrack_temp/results_for_hlca_datasets_HGNC Symbol_20221101-175029.pk"
)

results_core.keys()
results_core["HLCA_txt"][:3]

len(results_idtrack)
results_idtrack.keys()
results_idtrack["Kaminski_2020"].keys()
len(results_idtrack["Kaminski_2020"]["input_identifiers"])
results_idtrack["Kaminski_2020"]["input_identifiers"][:3]

### Pre-processing <a class="anchor" id="bullet3"></a>

for i in results_core.keys():
    ds = i.replace("_txt", "")
    print(ds)
    results_idtrack[ds] = dict()
    results_idtrack[ds]["input_identifiers"] = results_core[i]

results_idtrack.keys()

ds_rename = {"MeyerNikolic_unpubl": "Meyer_Nikolic_2022"}
results_idtrack[list(ds_rename.values())[0]] = results_idtrack[
    list(ds_rename.keys())[0]
]
del results_idtrack[list(ds_rename.keys())[0]]
results_idtrack.keys()

#### Feature names standardisation

final_simple = dict()
temp = list(results_idtrack.keys())
for ds in temp:
    ds_n = len(results_idtrack[ds]["input_identifiers"])
    print(f"\033[0;34m====================\033[0m {ds_n} | {ds}")
    ds_dict = {"original": [], "ensembl_id": [], "new": []}
    for i in results_idtrack[ds]["input_identifiers"]:
        # assert i["query_id"] not in final_simple[ds]
        ensembl_id = ""
        symbol = ""
        if len(i["last_node"]) > 0:
            ensembl_id = i["last_node"][0][0]
        if len(i["target_id"]) > 0:
            if not re.match(r"^ENSG", i["target_id"][0]):
                symbol = i["target_id"][0]
            else:
                symbol = ensembl_id
        else:
            symbol = i["query_id"]
        ds_dict["original"].append(i["query_id"])
        ds_dict["ensembl_id"].append(ensembl_id)
        ds_dict["new"].append(symbol)
    repeats = {x: count for x, count in Counter(ds_dict["new"]).items() if count >= 2}
    repeats = list(repeats.keys())
    final_simple[ds] = pd.DataFrame(
        {
            "original": ds_dict["original"],
            "ensembl_id": ds_dict["ensembl_id"],
            "new": ds_dict["new"],
            "repeat": [i in repeats for i in ds_dict["new"]],
        }
    )

print(ds)
temp = (
    final_simple[ds]
    .loc[~final_simple[ds]["ensembl_id"].isin([""]), :]
    .loc[~final_simple[ds]["new"].isin([""]), :]
    .loc[final_simple[ds]["repeat"], :]
    .sort_values(by=["new"])
)
temp

feature_old = {}
for ds in results_idtrack.keys():
    print(f"\033[94m*******************\033[0m {ds}")
    with open(f"{feats_dir}/{ds}.txt") as f:
        feature_old[ds] = [line[:-1] for line in f]
    temp = list(feature_old[ds])
    tvar = list(final_simple[ds]["ensembl_id"])
    brie = list(final_simple[ds]["original"])
    ephe = list(final_simple[ds]["new"])
    print(f"({len(temp)}) read:", temp[:3])
    print(f"({len(tvar)}) ense:", tvar[:3])
    print(f"({len(brie)}) orig:", brie[:3])
    print(f"({len(ephe)}) symb:", ephe[:3])

### Main <a class="anchor" id="bullet4"></a>

Expand shared ones

feature_conversion = final_simple.copy()
for ds_merged in adata_split.keys():
    print(f"\033[1m\033[93mExpanding\033[0m {ds_merged}")
    ds_column = list(adata_split[ds_merged].keys())[0]
    for ds in adata_split[ds_merged][ds_column]:
        feature_conversion[ds] = feature_conversion[ds_merged]
    del feature_conversion[ds_merged]

Concatenate to pivot create a table with studies as columns.

temp = list(final_simple.keys())
feature_conversion_all = final_simple[temp[0]].copy()
feature_conversion_all["study"] = temp[0]
for ds in temp[1:]:
    tvar = final_simple[ds].copy()
    tvar["study"] = ds.replace("HLCA", "Core")
    feature_conversion_all = pd.concat(
        [feature_conversion_all, tvar], axis=0, ignore_index=True
    )
feature_conversion_all

temp = len(set(feature_conversion_all["new"].tolist()))
print(f"Total: {temp}")

temp = feature_conversion_all.loc[
    feature_conversion_all["new"].isin(["HLA-B"]), :
]  # GTF2H2
temp.sort_values(by=["study"])

artifacts = [" " in i or "\n" in i for i in feature_conversion_all["new"]]
feature_conversion_all.loc[artifacts, :]

feature_conversion_all.loc[artifacts & feature_conversion_all["repeat"], :].sort_values(
    by=["new", "study"]
)

feats_out_table = re.sub("-with-core.pk", "_table_flat.csv", feats_inp)
feats_out_table

feature_conversion_all.to_csv(feats_out_table)

%%time
feature_conversion_all["old_aggr"] = feature_conversion_all.groupby(["new", "study"])[
    "original"
].transform(lambda x: ";".join(x))

feature_conversion_all.loc[feature_conversion_all["repeat"], :].sort_values(
    by=["new", "study"]
)

df = feature_conversion_all.drop_duplicates(subset=["new", "study"])
df.shape

df.loc[df["repeat"], :].sort_values(by=["new", "study"])

feature_conversion_df = df.pivot(
    index="new", columns="study", values="old_aggr"
).reset_index()
feature_conversion_df.shape
feature_conversion_df

feature_conversion_df.loc[feature_conversion_df["new"].isin(["HLA-B"]), :]

### Conclusions <a class="anchor" id="bullet5"></a>

### Save <a class="anchor" id="bullet6"></a>

feats_out_table = re.sub("-with-core.pk", "_table.csv", feats_inp)
feats_out_table

feature_conversion_df.to_csv(feats_out_table)

feats_out_pk = re.sub("-with-core.pk", "_digested.pk", feats_inp)
feats_out_pk

with open(f"{feats_out_pk}", "wb") as handle:
    pickle.dump(feature_conversion, handle, protocol=pickle.HIGHEST_PROTOCOL)

Done.
