==> notebooks/3_atlas_extension/14_full-matrix_object.ipynb <==
## Full matrix data object

We are now completing the whole object.

__authors__ = ["Ciro Ramírez-Suástegui", "Lisa Sikkema"]

__copyright__ = "Copyright 2022-11-15, Helmholtz Zentrum Muenchen"

__license__ = "GPL"

__version__ = "0.0.9"

__email__ = "ciro.suastegui@helmholtz-muenchen.de, lisa.sikkema@helmholtz-muenchen.de"

__status__ = "Prototype"

#### Structure

* [Global variables and paths](#bullet1)
* [Loading data](#bullet2)
* [Pre-processing](#bullet3)
* [Main](#bullet4)
* [Conclusions](#bullet5)
* [Save](#bullet6)

### Environment setup

%load_ext autoreload
%autoreload 2
import importlib

spam_spec = importlib.util.find_spec("lab_black")
if spam_spec is not None:
    %load_ext lab_black

# basic modules
import warnings, os, re
import time, sys, gc

warnings.simplefilter(action="ignore", category=FutureWarning)
warnings.simplefilter(action="ignore", category=UserWarning)

# in-house/developing modules
# tools modules
import scanpy as sc
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import scipy.sparse as sp

sc.logging.print_versions()

print("Environment:", re.sub(".os.py", "", os.__file__))
parentpath = lambda _path, n: os.sep.join(_path.split(os.sep)[:-n])
print("Working at:", re.sub(parentpath(os.getcwd(), 2), "", os.getcwd()))

### Global variables and paths <a class="anchor" id="bullet1"></a>

#### Inputs

# .layers: counts
cdata_inp = "../../data/HLCA_extended/HLCA_extended/HLCA_extended_full_raw.h5ad"
# .layers: soupXed
sdata_inp = "../../data/HLCA_extended/HLCA_extended/HLCA_extended_full.h5ad"
# .var
gname_inp = (
    "../../data/HLCA_extended/extension_datasets/features/all_update_table_flat.csv"
)
# uns: 'neighbors'
# obsm: 'X_umap'
# obsp: 'connectivities', 'distances'
# obsm: 'X_scanvi_emb' is in .X
embed_inp = "../../data/HLCA_extended/HLCA_extended/HLCA_extended_scarches_emb.h5ad"

#### Output

adata_out = "../../data/HLCA_extended/HLCA_extended/HLCA_extended_full_object.h5ad"

### Loading data <a class="anchor" id="bullet2"></a>

%%time
sdata = sc.read(sdata_inp, backed='r')

sdata

%%time
cdata = sc.read(cdata_inp, backed='r')

cdata

%%time
embed = sc.read(embed_inp)

embed

gname = pd.read_csv(gname_inp, index_col=0)

gname

### Pre-processing <a class="anchor" id="bullet3"></a>

#### Setting up features' metadata

%%time
gname["original_aggr"] = gname.groupby(["new"])["original"].transform(
    lambda x: ";".join(list(set(x)))
)

df = gname.copy()
df = df.loc[:, ["original_aggr", "ensembl_id", "new"]]

df = df.drop_duplicates(subset=["new"]).set_index("new")
df.shape
df

ensembl_id_duplicates = list(
    set(df.loc[df["ensembl_id"].duplicated(), "ensembl_id"]) - set(["", np.nan])
)
ensembl_id_duplicates[:5]

if len(ensembl_id_duplicates) > 0:
    temp = df.loc[df["ensembl_id"].isin(ensembl_id_duplicates), :].sort_values(
        by=["ensembl_id"], ascending=False
    )
    print(temp.iloc[:5, :])
else:
    print("No repeats in ensembl IDs")

df.loc[sdata.var_names, :]

#### Make sure `obs_names` have the same order

print(f"Same values? {set(sdata.obs_names) == set(embed.obs_names)}")

print(f"Same order? {(sdata.obs_names == embed.obs_names).all()}")

embed_obs = embed.obs_names.tolist()
meyer_obs = sdata.obs_names[sdata.obs["study"] == "Meyer_2021"].tolist()
other_obs = sdata.obs_names[sdata.obs["study"] != "Meyer_2021"].tolist()

Sorting embedding to the same order as the final Meyer + Rest (zero's only) matrix

embed = embed[meyer_obs + other_obs]

### Main <a class="anchor" id="bullet4"></a>

%%time
adata_c = sdata[meyer_obs + other_obs].to_memory()

adata_c.var = df.loc[adata_c.var_names, :]

adata_c.uns["neighbors"] = embed.uns["neighbors"]

adata_c.obsp["connectivities"] = embed.obsp["connectivities"]

adata_c.obsp["distances"] = embed.obsp["distances"]

adata_c.obsm["X_umap"] = embed.obsm["X_umap"]

Do you know if I need to use `X.copy()` or `.X` is fine?

Putting the embedding into `obsm`.

adata_c.obsm["X_scanvi_emb"] = embed.X.copy()

del embed

#### Normalization and log-transformation

%%time
sc.pp.normalize_total(adata_c, target_sum=7699)

%%time
sc.pp.log1p(adata_c)

#### Save raw counts in a layer and calculate total counts

We reload this to memory because we wanted to avoid it taking up memory while normalizing and log-transforming

%%time
adata_c.layers["counts"] = cdata[meyer_obs + other_obs].to_memory().X.copy()

adata_c.obs["total_counts"] = adata_c.layers["counts"].sum(axis=1)

#### Adding Meyer_2021 soupXed data

%%time
sdata_meyer2021 = sdata[meyer_obs].to_memory().X.copy()

adata_c.layers["soupX"] = sp.vstack(
    (
        sdata_meyer2021,
        sp.csr_matrix(
            (sdata.shape[0] - sdata_meyer2021.shape[0], sdata.shape[1]),
            dtype=np.float32,
        ),
    ),
    format="csr",
)

#### Checking matrices' sparcity

adata_c.X

adata_c.layers["counts"]

adata_c.layers["soupX"]

Meyer specifically

adata_c[adata_c.obs["study"] == "Meyer_2021"].X

adata_c[adata_c.obs["study"] == "Meyer_2021"].layers["counts"]

adata_c[adata_c.obs["study"] == "Meyer_2021"].layers["soupX"]

Tata because other studies get jealous

adata_c[adata_c.obs["study"] == "Tata_unpubl"].X

adata_c[adata_c.obs["study"] == "Tata_unpubl"].layers["counts"]

adata_c[adata_c.obs["study"] == "Tata_unpubl"].layers["soupX"]

Closing up to Meyer counts difference across matrices

adata_c[adata_c.obs["study"] == "Meyer_2021"].X[-10:, -10:-5].toarray()

adata_c[adata_c.obs["study"] == "Meyer_2021"].layers["counts"][-10:, -10:-5].toarray()

adata_c[adata_c.obs["study"] == "Meyer_2021"].layers["soupX"][-10:, -10:-5].toarray()

Checking other parts of the matrix

adata_c[adata_c.obs["study"] == "Meyer_2021"].X[-210:-200, -905:-900].toarray()

adata_c[adata_c.obs["study"] == "Meyer_2021"].layers["counts"][
    -210:-200, -905:-900
].toarray()

adata_c[adata_c.obs["study"] == "Meyer_2021"].layers["soupX"][
    -210:-200, -905:-900
].toarray()

Checking if Kaminski's data is integers.

adata_c[adata_c.obs["study"] == "Kaminski_2020"].X

adata_c[adata_c.obs["study"] == "Kaminski_2020"].layers["counts"]

adata_c[adata_c.obs["study"] == "Kaminski_2020"].layers["soupX"]

adata_c[adata_c.obs["study"] == "Kaminski_2020"].layers["counts"][
    -10:, -10:-5
].toarray()

pd.Series(
    adata_c[adata_c.obs["study"] == "Kaminski_2020"]
    .layers["counts"][:100, :]
    .A.flatten()
).value_counts()

%%time
non_int = adata_c[adata_c.obs["study"] == "Kaminski_2020"].layers["counts"].toarray() % 1 != 0

total_non_ints = sum(non_int)

sum(total_non_ints)

### Conclusions <a class="anchor" id="bullet5"></a>

adata_c

### Save <a class="anchor" id="bullet6"></a>

%%time
adata_c.write(filename=adata_out)

Done.
